Added «Quebec» as a language in the server-side Java
Added Québec as a designator instead of French because of important cultural differences between France, Québec, and the rest of the francophony. Mostly Québec names are more often double barrelled, and so are our last names. Utlimately, a designation as fr-ca or fr-qc would certainly be fine, I just don't want to introduce a bothersome conflation of name groups.